Split assert module
