+++ py-polars/polars/testing/__init__.py
@@ -6,8 +6,8 @@ from polars.testing.asserts import (
 )
 
 __all__ = [
-    "assert_series_equal",
-    "assert_series_not_equal",
     "assert_frame_equal",
     "assert_frame_not_equal",
+    "assert_series_equal",
+    "assert_series_not_equal",
 ]
