+++ py-polars/polars/testing/asserts/__init__.py
@@ -1,0 +1,9 @@
+from polars.testing.asserts.frame import assert_frame_equal, assert_frame_not_equal
+from polars.testing.asserts.series import assert_series_equal, assert_series_not_equal
+
+__all__ = [
+    "assert_frame_equal",
+    "assert_frame_not_equal",
+    "assert_series_equal",
+    "assert_series_not_equal",
+]
