+++ py-polars/polars/testing/asserts/utils.py
@@ -1,0 +1,15 @@
+from __future__ import annotations
+
+from typing import Any, NoReturn
+
+
+def raise_assertion_error(
+    obj: str,
+    detail: str,
+    left: Any,
+    right: Any,
+) -> NoReturn:
+    """Raise a detailed assertion error."""
+    __tracebackhide__ = True
+    msg = f"{obj} are different ({detail})\n[left]:  {left}\n[right]: {right}"
+    raise AssertionError(msg)
